Gateesaenew/Gate/requirements.txt
Django==4.2.0
djangorestframework
xhtml2pdf
firebase-admin
Pillow
